Creando mensajes de error

+++ SaldoVacaciones/Models/ErrorModel.cs
@@ -3,7 +3,7 @@
     public class ErrorModel
     {
         public int Id { get; set; }
-        public short Codigo { get; set; }
+        public int Codigo { get; set; }
         public string? Descripcion { get; set; }
 
     }
